fix: move git url handling to bootstrapper
Move source resolution for requirements with git URLs into the
bootstrapper. This will allow us to ensure we have the build
dependencies for a package built and installed before we try to get
its other metadata.
