+++ e2e/test_bootstrap_git_url_tag.sh
@@ -6,9 +6,10 @@
 SCRIPTDIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 source "$SCRIPTDIR/common.sh"
 
-GIT_REPO_URL="https://opendev.org/openstack/stevedore.git"
+GIT_REPO_URL="https://github.com/python-wheel-build/stevedore-test-repo.git"
 
 fromager \
+  --verbose \
   --debug \
   --log-file="$OUTDIR/bootstrap.log" \
   --error-log-file="$OUTDIR/fromager-errors.log" \
